requirements.txt
python-dotenv==1.0.0
boto3==1.34.29
PyGithub==2.1.1
openai==1.12.0
redis==5.0.1
anthropic==0.45.0